Refactor imports to use _types from core module; update related tests and remove deprecated typings references.
