+++ opticalib/__init__.py
@@ -31,6 +31,7 @@ from .core.root import (
     set_configuration_file,
 )
 from .core import config
+from .core import _types as typings
 from .core.fitsarray import fits_array
 from .devices import *
 from .devices.interferometer import _4DInterferometer
@@ -66,6 +67,7 @@ __all__ = [
     "create_configuration_file",
     "set_configuration_file",
     "config",
+    "typings",
     "get_camera_settings",
     "get_frame_rate",
     "fits_array",
