+++ opticalib/devices/_API/alpaoAPI.py
@@ -1,7 +1,7 @@
 import numpy as _np
 from opticalib.core.config import get_section_config
 from opticalib.core.exceptions import CommandError
-from opticalib import typings as _t
+from opticalib.core import _types as _t
 
 
 class BaseAlpaoMirror:
