+++ opticalib/devices/_API/piAPI.py
@@ -1,5 +1,5 @@
 import numpy as np
-from opticalib import typings as _ot
+from opticalib.core import _types as _ot
 from pipython import GCSDevice, GCSError
 from pipython.pidevice.interfaces.pisocket import PISocket
 from opticalib.core import config as _rc
