+++ tests/test_measurements.py
@@ -7,7 +7,7 @@ import os
 import numpy as np
 import numpy.ma as ma
 from unittest.mock import MagicMock, Mock, patch
-from opticalib.measurements import Measurements
+from opticalib.procedures import Measurements
 
 
 class _FakeInterferometer:
@@ -111,8 +111,8 @@ class TestAcquireTimeSeries:
         data_subfolder = os.path.join(opd_folder, "20260403_080000")
         os.makedirs(data_subfolder, exist_ok=True)
 
-        monkeypatch.setattr("opticalib.measurements._cdf", lambda path: data_subfolder)
-        monkeypatch.setattr("opticalib.measurements._save_fits", lambda path, data: None)
+        monkeypatch.setattr("opticalib.procedures.measurements._cdf", lambda path: data_subfolder)
+        monkeypatch.setattr("opticalib.procedures.measurements._save_fits", lambda path, data: None)
 
         result = m.acquire_time_series(nframes=2)
 
@@ -134,8 +134,8 @@ class TestAcquireTimeSeries:
 
         data_subfolder = os.path.join(temp_dir, "sub")
         os.makedirs(data_subfolder, exist_ok=True)
-        monkeypatch.setattr("opticalib.measurements._cdf", lambda path: data_subfolder)
-        monkeypatch.setattr("opticalib.measurements._save_fits", lambda path, data: None)
+        monkeypatch.setattr("opticalib.procedures.measurements._cdf", lambda path: data_subfolder)
+        monkeypatch.setattr("opticalib.procedures.measurements._save_fits", lambda path, data: None)
 
         nframes = 3
         m.acquire_time_series(nframes=nframes)
@@ -158,8 +158,8 @@ class TestAcquireTimeSeries:
 
         data_subfolder = os.path.join(temp_dir, "sub")
         os.makedirs(data_subfolder, exist_ok=True)
-        monkeypatch.setattr("opticalib.measurements._cdf", lambda path: data_subfolder)
-        monkeypatch.setattr("opticalib.measurements._save_fits", lambda path, data: None)
+        monkeypatch.setattr("opticalib.procedures.measurements._cdf", lambda path: data_subfolder)
+        monkeypatch.setattr("opticalib.procedures.measurements._save_fits", lambda path, data: None)
 
         nframes = 2
         m.acquire_time_series(nframes=nframes)
